+++ csvviz/cmds/area.py
@@ -15,33 +15,41 @@ from csvviz.vizkit import Vizkit
 class Areakit(Vizkit):
     viz_type = "area"
     viz_info = f"""An area chart. Can be stacked"""
+    viz_epilog = """Example:  $ csvviz area -x date -y price -c company stocks.csv"""
 
     COMMAND_DECORATORS = (
         click.option(
-            "--xvar", "-x", type=click.STRING, default="", help="the label column"
+            "--xvar",
+            "-x",
+            type=click.STRING,
+            help="The name of the column for mapping x-axis values; if empty, the first (columns[0]) column is used",
         ),
         click.option(
-            "--yvar", "-y", type=click.STRING, default="", help="the value column"
+            "--yvar",
+            "-y",
+            type=click.STRING,
+            help="The name of the column for mapping y-axis values; if empty, the second (columns[1]) column is used",
         ),
         click.option(
             "--color",
             "-c",
             "fillvar",
             type=click.STRING,
-            help="The column used to specify fill color",
+            help="The name of the column for mapping bar colors. This is required for creating a stacked chart.",
         ),
         # https://altair-viz.github.io/user_guide/encoding.html?#ordering-marks
         click.option(
             "--color-sort",
             "-cs",
             "fillsort",
-            type=click.Choice(("+", "-")),
-            help="Whether to sort the fill stack in ascending (+) or descending (-) order by nominal name TKTK",
+            type=click.Choice(("asc", "desc"), case_sensitive=False),
+            help="For stacked charts, the sort order of the color variable: 'asc' for ascending, 'desc' for descending/reverse",
         ),
     )
 
     def prepare_channels(self):
         channels = self._create_channels(self.channel_kwargs)
+        self._set_channel_colorscale("fill", channels)
 
         # sort by fill/stack is not the same as sorting the x-axis:
         # https://altair-viz.github.io/user_guide/encoding.html?#ordering-marks
@@ -52,7 +60,7 @@ class Areakit(Vizkit):
                 )
             else:
                 fname = self.resolve_channel_name(channels["fill"])
-                fsort = "descending" if _fillsort == "-" else "ascending"
+                fsort = "descending" if _fillsort == "desc" else "ascending"
                 channels["order"] = alt.Order(fname, sort=fsort)
 
         return channels