+++ csvviz/cmds/hist.py
@@ -9,46 +9,91 @@ import altair as alt
 import click
 
 from csvviz.exceptions import *
-from csvviz.vizkit import Vizkit
+from csvviz.cmds.bar import Barkit
+from csvviz.cli_utils import clerr
+
+BINNING_OPTS = (
+    "bincount",
+    "binstepsize",
+)
 
 
-class Histkit(Vizkit):
+class Histkit(Barkit):
     viz_type = "hist"
-    viz_info = f"""An bar chart, showing counts/bins TK"""
+    viz_info = f"""A bar chart that maps the frequency count of a given variable. Can be stacked."""
+    viz_epilog = """Example:\t csvviz hist -x Horsepower examples/cars.csv"""
 
     def prepare_channels(self):
-        channels = self._create_channels(self.channel_kwargs)
 
-        ##### bin time
-        channels["x"].bin = True
-        channels["y"] = "count()"
-        #####
+        channels = super().prepare_channels()
+        bwargs = {k: self.kwargs[k] for k in BINNING_OPTS if self.kwargs.get(k)}
 
-        if self.kwargs.get("flipxy"):  # i.e. -H/--horizontal flag
-            channels["x"], channels["y"] = (channels["y"], channels["x"])
+        # deal with special case in which xvar is a nominal field, which means user
+        # is trying to do a standard frequency count
+        if channels["x"].type == "nominal":
+            xname = self.resolve_channel_name(channels["x"])
+            channels["y"] = {"aggregate": "count", "field": xname, "type": "nominal"}
 
-        if channels.get("fill"):
-            channels["fill"].scale = alt.Scale(**self._config_colors(self.color_kwargs))
-            channels["fill"].legend = None
+            # issue warning if any binning options were set:
+            if bwargs:
+                self.warnings.append(
+                    f"""Since '{xname}' consists of nominal values, csvviz will ignore bin-specific settings, e.g. -n/--bins and -s/--bin-size"""
+                )
+
+        else:
+            channels[
+                "y"
+            ] = "count()"  # explicitly set/override y to always be an aggregate count
+            # if no bin opts are set, then encoding.x.bins is just True
+            #   and Vega defaults are used
+            #   https://vega.github.io/vega-lite/docs/bin.html#bin-parameters
+            channels["x"].bin = True
+            if bwargs:
+                bdict = {}
+                if _n := bwargs.get("bincount"):
+                    bdict["maxbins"] = _n
+
+                if _s := bwargs.get("binstepsize"):
+                    bdict["step"] = _s
+                    bdict.pop("maxbins", None)  # override maxbins
+
+                channels["x"].bin = alt.Bin(**bdict)
 
         return channels
 
     COMMAND_DECORATORS = (
         click.option(
-            "--xvar", "-x", type=click.STRING, default="", help="the thing to bin TK"
+            "--xvar",
+            "-x",
+            type=click.STRING,
+            help="The name of the column for mapping frequency count to the x-axis",
         ),
         click.option(
-            "--fill",
-            "-f",
+            "--color",
+            "-c",
             "fillvar",
             type=click.STRING,
-            help="The column used to specify fill color",
+            help="The name of the column for mapping bar colors. This is required for creating a stacked chart.",
         ),
         click.option(
             "--horizontal",
             "-H",
             "flipxy",
             is_flag=True,
-            help="Orient the bars horizontally",
+            help="Make a horizontal bar chart",
+        ),
+        click.option(
+            "--bins",
+            "-n",
+            "bincount",
+            type=click.INT,
+            help="Specify a max number of bins (overridden by -s/--bin-size)",
+        ),
+        click.option(
+            "--bin-size",
+            "-s",
+            "binstepsize",
+            type=click.FLOAT,
+            help="Specify a size for each bin (overrides -n/--bins)",
         ),
     )