+++ csvviz/cmds/scatter.py
@@ -12,34 +12,39 @@ from csvviz.vizkit import Vizkit
 
 class Scatterkit(Vizkit):
     viz_type = "scatter"
-    viz_info = f"""A scatterplot for showing relationship between two independent variables x and y. Can vary by size and fill too. TK"""
+    viz_info = f"""A scatterplot for showing relationship between two independent variables x and y. Set -s/--size to create a bubble (variable dot size) chart"""
+    viz_epilog = """Example:  $ csvviz scatter -x mass -y volume -s velocity data.csv"""
 
     def prepare_channels(self):
         channels = self._create_channels(self.channel_kwargs)
-        if channels.get("fill"):
-            channels["fill"].scale = alt.Scale(**self._config_colors(self.color_kwargs))
-
+        self._set_channel_colorscale("fill", channels)
         return channels
 
     COMMAND_DECORATORS = (
         click.option(
-            "--xvar", "-x", type=click.STRING, default="", help="the label column"
+            "--xvar",
+            "-x",
+            type=click.STRING,
+            help="The name of the column for mapping x-axis values; if empty, the first (columns[0]) column is used",
         ),
         click.option(
-            "--yvar", "-y", type=click.STRING, default="", help="the value column"
+            "--yvar",
+            "-y",
+            type=click.STRING,
+            help="The name of the column for mapping y-axis values; if empty, the second (columns[1]) column is used",
         ),
         click.option(
-            "--fill",
-            "-f",
+            "--color",
+            "-c",
             "fillvar",
             type=click.STRING,
-            help="The column used to specify fill color",
+            help="The name of the column for mapping dot colors. This is required for creating a multi-series scatter chart.",
         ),
         click.option(
             "--size",
             "-s",
             "sizevar",
             type=click.STRING,
-            help="The column used to specify dot size",
+            help="The name of the column for mapping dot size. This is required for creating a bubble chart.",
         ),
     )